+++ src/components/Dropdown/Dropdown.types.ts
@@ -1,6 +1,20 @@
 import React, { Ref } from 'react';
 import { Placement, Strategy } from '@floating-ui/react';
 
+export const ANIMATION_DURATION: number = 200;
+export const NO_ANIMATION_DURATION: number = 10;
+export const PREVENT_DEFAULT_TRIGGERS: string[] = ['contextmenu'];
+export const TRIGGER_TO_HANDLER_MAP_ON_ENTER = {
+  click: 'onClick',
+  hover: 'onMouseEnter',
+  contextmenu: 'onContextMenu',
+};
+export const TRIGGER_TO_HANDLER_MAP_ON_LEAVE = {
+  click: '',
+  hover: 'onMouseLeave',
+  contextmenu: '',
+};
+
 export interface DropdownProps {
   /**
    * Class names of the main wrapper
@@ -11,6 +25,11 @@ export interface DropdownProps {
    * @default true
    */
   closeOnDropdownClick?: boolean;
+  /**
+   * Should close Tooltip on reference click.
+   * @default true
+   */
+  closeOnReferenceClick?: boolean;
   /**
    * Should close dropdown on click outside
    * @default true
@@ -62,7 +81,7 @@ export interface DropdownProps {
    */
   portal?: boolean;
   /**
-   * Positioning strategy for the tooltip
+   * Positioning strategy for the dropdown
    * @default absolute
    */
   positionStrategy?: Strategy;
@@ -88,6 +107,10 @@ export interface DropdownProps {
    * Style of the main wrapper
    */
   style?: React.CSSProperties;
+  /**
+   * The optional tab index of the reference element.
+   */
+  tabIndex?: number;
   /**
    * The trigger mode that opens the dropdown
    * @default 'click'